Reorganize Charts module into utils, types, and hooks folders.
Move chart helpers out of context so providers only own React context wiring.
