+++ src/components/Charts/context/ChartDefaultTypefaceProvider.native.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
+import useChartSkiaTypefaces from '../hooks/useChartSkiaTypefaces';
 import {ChartDefaultTypefaceContext} from './ChartDefaultTypefaceContext';
-import useChartSkiaTypefaces from './useChartSkiaTypefaces';
 
 function ChartDefaultTypefaceProvider({children}: {children: React.ReactNode}) {
     const typefaces = useChartSkiaTypefaces();
