+++ src/components/Charts/hooks/index.ts
@@ -4,7 +4,7 @@ export {default as useChartParagraphs} from './useChartParagraphs';
 export {default as useYAxisLabelWidth} from './useYAxisLabelWidth';
 export {default as useChartFontManager} from './useChartFontManager/useChartFontManager';
 export {default as ChartDefaultTypefaceProvider} from '../context/ChartDefaultTypefaceProvider';
-export {getChartSkiaTypeface} from '../context/getChartSkiaTypeface';
+export {getChartSkiaTypeface} from '../utils/getChartSkiaTypeface';
 export {useChartDefaultTypeface} from '../context/ChartDefaultTypefaceContext';
 export {useChartInteractions, TOOLTIP_BAR_GAP} from './useChartInteractions';
 export type {HitTestArgs} from './useChartInteractions';
